+++ pkgs/dart_mcp/lib/src/api/prompts.dart
@@ -86,7 +86,8 @@ extension type GetPromptResult.fromMap(Map<String, Object?> _value)
 }
 
 /// A prompt or prompt template that the server offers.
-extension type Prompt.fromMap(Map<String, Object?> _value) {
+extension type Prompt.fromMap(Map<String, Object?> _value)
+    implements BaseMetadata {
   factory Prompt({
     required String name,
     String? description,
@@ -97,9 +98,6 @@ extension type Prompt.fromMap(Map<String, Object?> _value) {
     if (arguments != null) 'arguments': arguments,
   });
 
-  /// The name of the prompt or prompt template.
-  String get name => _value['name'] as String;
-
   /// An optional description of what this prompt provides.
   String? get description => _value['description'] as String?;
 
@@ -108,19 +106,19 @@ extension type Prompt.fromMap(Map<String, Object?> _value) {
 }
 
 /// Describes an argument that a prompt can accept.
-extension type PromptArgument.fromMap(Map<String, Object?> _value) {
+extension type PromptArgument.fromMap(Map<String, Object?> _value)
+    implements BaseMetadata {
   factory PromptArgument({
     required String name,
+    String? title,
     String? description,
     bool? required,
   }) => PromptArgument.fromMap({
     'name': name,
+    if (title != null) 'title': title,
     if (description != null) 'description': description,
     if (required != null) 'required': required,
   });
-
-  /// The name of the argument.
-  String get name => _value['name'] as String;
 
   /// A human-readable description of the argument.
   String? get description => _value['description'] as String?;